configs: New config in drivers/crypto for v5.4-rc1
Message-id: <20191014162300.16308-4-jcline@redhat.com>
Patchwork-id: 278888
O-Subject: [ARK INTERNAL PATCH] configs: New config in drivers/crypto for v5.4-rc1
Bugzilla:
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/crypto/ccp/Kconfig (help): CRYPTO_DEV_CCP_DEBUGFS (bool)
    Expose CCP device information such as operation statistics, feature
    information, and descriptor queue contents.
* drivers/crypto/hisilicon/Kconfig (help): CRYPTO_DEV_HISI_ZIP (tristate)
    Support for HiSilicon ZIP Driver
* drivers/crypto/Kconfig (help): CRYPTO_DEV_SAFEXCEL (tristate)
    This driver interfaces with the SafeXcel EIP-97 and EIP-197 cryptographic
    engines designed by Inside Secure. It currently accelerates DES, 3DES and
    AES block ciphers in ECB and CBC mode, as well as SHA1, SHA224, SHA256,
    SHA384 and SHA512 hash algorithms for both basic hash and HMAC.
    Additionally, it accelerates combined AES-CBC/HMAC-SHA AEAD operations.

Cc: Neil Horman <nhorman@redhat.com>
Cc: Michael S. Tsirkin <mst@redhat.com>
Cc: Jason Wang <jasowang@redhat.com>
Cc: rhvirt-patches@redhat.com
Signed-off-by: Jeremy Cline <jcline@redhat.com>