Introduce rh-configs-commit
One of the tasks that comes with maintaining ARK is dealing with new
configuration options that come in. We need to set an initial value
for building a kernel and then allow for a final review of the options.
This has been managed by out of tree scripts for a while but to make
automation easier, bring part of this in tree.

Introduce the rh-configs-commit make target. This runs 'make helpnewconfig'
(like listnewconfig execpt it shows the help text) and puts all
the new configuration options in the pending directory. Moving those
options out of the pending directory will be a separate step.

Signed-off-by: Laura Abbott <labbott@redhat.com>