[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ARCH_S32:

 This enables support for the NXP S32 family of processors.

 Symbol: ARCH_S32 [=n]
 Type  : bool
 Prompt: NXP S32 SoC Family
   Location:
     -> Platform selection
   Defined at arch/arm64/Kconfig.platforms:217

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>