[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_LEDS_EL15203000:

 This option enables support for EL15203000 LED Board
 (aka RED LED board) which is widely used in coffee vending
 machines produced by Crane Merchandising Systems.

 To compile this driver as a module, choose M here: the module
 will be called leds-el15203000.

 Symbol: LEDS_EL15203000 [=n]
 Type  : tristate
 Prompt: LED Support for Crane EL15203000
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
   Defined at drivers/leds/Kconfig:135
   Depends on: NEW_LEDS [=y] && LEDS_CLASS [=y] && SPI [=y] && OF [=y]

---