[redhat] New configs in drivers/mmc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MMC_SDHCI_MILBEAUT:

 This selects the Secure Digital Host Controller Interface (SDHCI)
 Needed by Milbeaut SoC for MMC / SD / SDIO support.
 If you have a controller with this interface, say Y or M here.

 If unsure, say N.

 Symbol: MMC_SDHCI_MILBEAUT [=n]
 Type  : tristate
 Prompt: SDHCI support for Socionext Milbeaut Serieas using F_SDH30
   Location:
     -> Device Drivers
       -> MMC/SD/SDIO card support (MMC [=m])
   Defined at drivers/mmc/host/Kconfig:372
   Depends on: MMC [=m] && MMC_SDHCI_PLTFM [=m] && OF [=y]

---