[redhat] New configs in lib/Kconfig.kasan
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_KASAN_VMALLOC:

 By default, the shadow region for vmalloc space is the read-only
 zero page. This means that KASAN cannot detect errors involving
 vmalloc space.

 Enabling this option will hook in to vmap/vmalloc and back those
 mappings with real shadow memory allocated on demand. This allows
 for KASAN to detect more sorts of errors (and to support vmapped
 stacks), but at the cost of higher memory usage.

 Symbol: KASAN_VMALLOC [=n]
 Type  : bool
 Prompt: Back mappings in vmalloc space with real shadow memory
   Location:
     -> Kernel hacking
       -> Memory Debugging
   Defined at lib/Kconfig.kasan:148
   Depends on: KASAN [=y] && HAVE_ARCH_KASAN_VMALLOC [=y]

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Laura Abbott <labbott@redhat.com>
Cc: Marcelo Ricardo Leitner <mleitner@redhat.com>