New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_QCOM_OCMEM:

 The On Chip Memory (OCMEM) allocator allows various clients to
 allocate memory from OCMEM based on performance, latency and power
 requirements. This is typically used by the GPU, camera/video, and
 audio components on some Snapdragon SoCs.

 Symbol: QCOM_OCMEM [=n]
 Type  : tristate
 Prompt: Qualcomm On Chip Memory (OCMEM) driver
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Qualcomm SoC drivers
   Defined at drivers/soc/qcom/Kconfig:69
   Depends on: ARCH_QCOM [=y]
   Selects: QCOM_SCM [=n]