[redhat] New configs in net/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ETHTOOL_NETLINK:

 An alternative userspace interface for ethtool based on generic
 netlink. It provides better extensibility and some new features,
 e.g. notification messages.

 Symbol: ETHTOOL_NETLINK [=y]
 Type  : bool
 Prompt: Netlink interface for ethtool
   Location:
     -> Networking support (NET [=y])
   Defined at net/Kconfig:453
   Depends on: NET [=y]

---

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Rashid Khan <rkhan@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Cc: Davide Caratti <dcaratti@redhat.com>
Cc: Eric Garver <egarver@redhat.com>
Cc: Flavio Leitner <fbl@redhat.com>
Cc: Florian Westphal <fwestpha@redhat.com>
Cc: Guillaume Nault <gnault@redhat.com>
Cc: Hangbin Liu <haliu@redhat.com>
Cc: Ivan Vecera <ivecera@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Cc: Marcelo Leitner <mleitner@redhat.com>
Cc: Matteo Croce <mcroce@redhat.com>
Cc: Paolo Abeni <pabeni@redhat.com>
Cc: Phil Sutter <psutter@redhat.com>
Cc: Sabrina Dubroca <sdubroca@redhat.com>
Cc: Stefano Brivio <sbrivio@redhat.com>
Cc: Xin Long <lxin@redhat.com>