[redhat] New configs in net/ipv4
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_INET_ESPINTCP:

 Support for RFC 8229 encapsulation of ESP and IKE over
 TCP/IPv4 sockets.

 If unsure, say N.

 Symbol: INET_ESPINTCP [=n]
 Type  : bool
 Prompt: IP: ESP in TCP encapsulation (RFC 8229)
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> TCP/IP networking (INET [=y])
           -> IP: ESP transformation (INET_ESP [=m])
   Defined at net/ipv4/Kconfig:381
   Depends on: NET [=y] && INET [=y] && XFRM [=y] && INET_ESP [=m]
   Selects: STREAM_PARSER [=y] && NET_SOCK_MSG [=y]

This has been enabled at the request of Marcelo Ricardo Leitner
<mleitner@redhat.com>
---

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Rashid Khan <rkhan@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Cc: Davide Caratti <dcaratti@redhat.com>
Cc: Eric Garver <egarver@redhat.com>
Cc: Flavio Leitner <fbl@redhat.com>
Cc: Florian Westphal <fwestpha@redhat.com>
Cc: Guillaume Nault <gnault@redhat.com>
Cc: Hangbin Liu <haliu@redhat.com>
Cc: Ivan Vecera <ivecera@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Cc: Marcelo Leitner <mleitner@redhat.com>
Cc: Matteo Croce <mcroce@redhat.com>
Cc: Paolo Abeni <pabeni@redhat.com>
Cc: Phil Sutter <psutter@redhat.com>
Cc: Sabrina Dubroca <sdubroca@redhat.com>
Cc: Stefano Brivio <sbrivio@redhat.com>
Cc: Xin Long <lxin@redhat.com>