[redhat] New configs in arch/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_STRICT_KERNEL_RWX:

 If this is set, kernel text and rodata memory will be made read-only,
 and non-text memory will be made non-executable. This provides
 protection against certain security exploits (e.g. executing the heap
 or modifying text)

 These features are considered standard security practice these days.
 You should say Y here in almost all cases.

 Symbol: STRICT_KERNEL_RWX [=n]
 Type  : bool
 Defined at arch/Kconfig:865
   Prompt: Make kernel text and rodata read-only
   Depends on: ARCH_HAS_STRICT_KERNEL_RWX [=y]
   Visible if: ARCH_HAS_STRICT_KERNEL_RWX [=y] && ARCH_OPTIONAL_KERNEL_RWX [=y]
   Location:
     -> General architecture-dependent options

---