[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MFD_ROHM_BD71828:

 Select this option to get support for the ROHM BD71828 Power
 Management IC. BD71828GW is a single-chip power management IC for
 battery-powered portable devices. The IC integrates 7 buck
 converters, 7 LDOs, and a 1500 mA single-cell linear charger.
 Also included is a Coulomb counter, a real-time clock (RTC), and
 a 32.768 kHz clock gate.

 Symbol: MFD_ROHM_BD71828 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:1910
   Prompt: ROHM BD71828 Power Management IC
   Depends on: HAS_IOMEM [=y] && I2C [=y]=y && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
 Selects: REGMAP_I2C [=m] && REGMAP_IRQ [=n] && MFD_CORE [=m]

---