[redhat] Dynamically set the debugbuildsenabled variable
ARK builds weekly, always based of a tag in Linus's tree. Fedora Rawhide
kernels are built daily and the policy is to use the debug
configurations except for RC and final releases.

redhat/Makefile.common determines if HEAD's merge base is a tag, and if
is, SNAPSHOT is set to 0. If HEAD's merge base is not a tag, SNAPSHOT is
1. Invert this and use it to dynamically set the debugbuildsenabled
boolean in the specfile.

Cc: Don Zickus <dzickus@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: Jan Stancek <jstancek@redhat.com>
Cc: Augusto Caringi <acaringi@redhat.com>
Cc: Frantisek Hrbata <fhrbata@redhat.com>
Cc: Luis Claudio Goncalves <lgoncalv@redhat.com>
Cc: Juri Lelli <jlelli@redhat.com>
Cc: Rado Vrbovsky <rvrbovsk@redhat.com>
Cc: Herton R. Krzesinski <herton@redhat.com>
Cc: Denys Vlasenko <dvlasenk@redhat.com>
Cc: Bruno Meneguele <bmeneg@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>