[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_APPLE_MFI_FASTCHARGE:

 Say Y here if you want to control whether iOS devices will
 fast charge from the USB interface, as implemented in "MFi"
 chargers.

 It is safe to say M here.

 Symbol: APPLE_MFI_FASTCHARGE [=n]
 Type  : tristate
 Defined at drivers/usb/misc/Kconfig:140
   Prompt: Fast charge control for iOS devices
   Depends on: USB_SUPPORT [=y] && USB [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
 Selects: POWER_SUPPLY [=y]

---

Cc: Torez Smith <torez@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>