New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DEFAULT_INIT:

 This option determines the default init for the system if no init=
 option is passed on the kernel command line. If the requested path is
 not present, we will still then move on to attempting further
 locations (e.g. /sbin/init, etc). If this is empty, we will just use
 the fallback list when init= is not passed.

 Symbol: DEFAULT_INIT [=]
 Type  : string
 Defined at init/Kconfig:267
   Prompt: Default init path
   Location:
     -> General setup

Cc: Prarit Bhargava <prarit@redhat.com>