New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PINCTRL_JASPERLAKE:

 This pinctrl driver provides an interface that allows configuring
 of Intel Jasper Lake PCH pins and using them as GPIOs.

 Symbol: PINCTRL_JASPERLAKE [=n]
 Type  : tristate
 Defined at drivers/pinctrl/intel/Kconfig:114
   Prompt: Intel Jasper Lake PCH pinctrl and GPIO driver
   Depends on: PINCTRL [=y] && (X86 [=y] || COMPILE_TEST [=n]) && ACPI [=y]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
 Selects: PINCTRL_INTEL [=m]

v2: enable for JASPERLAKE on x86

Cc: David Arcari <darcari@redhat.com>