New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SERIAL_LANTIQ:

 Support for UART on Lantiq and Intel SoCs.
 To compile this driver as a module, select M here. The
 module will be called lantiq.

 Symbol: SERIAL_LANTIQ [=n]
 Type  : tristate
 Defined at drivers/tty/serial/Kconfig:1036
   Prompt: Lantiq serial driver
   Depends on: TTY [=y] && HAS_IOMEM [=y] && (LANTIQ || X86 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
 Selects: SERIAL_CORE [=y]

Cc: John Linville <linville@redhat.com>