[redhat] New configs in drivers/mmc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MMC_SDHCI_OF_DWCMSHC:

 This selects Synopsys DesignWare Cores Mobile Storage Controller
 support.
 If you have a controller with this interface, say Y or M here.
 If unsure, say N.

 Symbol: MMC_SDHCI_OF_DWCMSHC [=n]
 Type  : tristate
 Defined at drivers/mmc/host/Kconfig:205
   Prompt: SDHCI OF support for the Synopsys DWC MSHC
   Depends on: MMC [=m] && MMC_SDHCI_PLTFM [=m] && OF [=y] && COMMON_CLK [=y]
   Location:
     -> Device Drivers
       -> MMC/SD/SDIO card support (MMC [=m])
         -> Secure Digital Host Controller Interface support (MMC_SDHCI [=m])
           -> SDHCI platform and OF driver helper (MMC_SDHCI_PLTFM [=m])

---

Cc: Gopal Tiwari <gtiwari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>