[redhat] New configs in drivers/nvme
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_NVME_TARGET_PASSTHRU:

 This enables target side NVMe passthru controller support for the
 NVMe Over Fabrics protocol. It allows for hosts to manage and
 directly access an actual NVMe controller residing on the target
 side, incuding executing Vendor Unique Commands.

 If unsure, say N.

 Symbol: NVME_TARGET_PASSTHRU [=n]
 Type  : bool
 Defined at drivers/nvme/target/Kconfig:19
   Prompt: NVMe Target Passthrough support
   Depends on: NVME_TARGET [=m] && (NVME_CORE [=m]=y || NVME_CORE [=m]=NVME_TARGET [=m])
   Location:
     -> Device Drivers
       -> NVME Support
         -> NVMe Target support (NVME_TARGET [=m])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>