[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_FAULT_INJECTION_USERCOPY:

 Provides fault-injection capability to inject failures
 in usercopy functions (copy_from_user(), get_user(), ...).

 Symbol: FAULT_INJECTION_USERCOPY [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:1771
   Prompt: Fault injection capability for usercopy functions
   Depends on: FAULT_INJECTION [=y]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Fault-injection framework (FAULT_INJECTION [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>