[redhat] New configs in lib/Kconfig.kgdb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_KGDB_HONOUR_BLOCKLIST:

 If set to Y the debug core will use the kprobe blocklist to
 identify symbols where it is unsafe to set breakpoints.
 In particular this disallows instrumentation of functions
 called during debug trap handling and thus makes it very
 difficult to inadvertently provoke recursive trap handling.

 If unsure, say Y.

 Symbol: KGDB_HONOUR_BLOCKLIST [=y]
 Type  : bool
 Defined at lib/Kconfig.kgdb:27
   Prompt: KGDB: use kprobe blocklist to prohibit unsafe breakpoints
   Depends on: KGDB [=y] && HAVE_KPROBES [=y] && MODULES [=y]
   Location:
     -> Kernel hacking
       -> Generic Kernel Debugging Instruments
         -> KGDB: kernel debugger (KGDB [=y])
 Selects: KPROBES [=y]

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>