[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CSD_LOCK_WAIT_DEBUG:

 This option enables debug prints when CPUs are slow to respond
 to the smp_call_function*() IPI wrappers.  These debug prints
 include the IPI handler function currently executing (if any)
 and relevant stack traces.

 Symbol: CSD_LOCK_WAIT_DEBUG [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:1380
   Prompt: Debugging for csd_lock_wait(), called from smp_call_function*()
   Depends on: DEBUG_KERNEL [=y] && 64BIT [=y]
   Location:
     -> Kernel hacking
       -> Lock Debugging (spinlocks, mutexes, etc...)

---

 CONFIG_SCF_TORTURE_TEST:

 This option provides a kernel module that runs torture tests
 on the smp_call_function() family of primitives.  The kernel
 module may be built after the fact on the running kernel to
 be tested, if desired.

 Symbol: SCF_TORTURE_TEST [=n]
 Type  : tristate
 Defined at lib/Kconfig.debug:1370
   Prompt: torture tests for smp_call_function*()
   Depends on: DEBUG_KERNEL [=y]
   Location:
     -> Kernel hacking
       -> Lock Debugging (spinlocks, mutexes, etc...)
 Selects: TORTURE_TEST [=n]

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>