[redhat] New configs in drivers/mailbox
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ARM_MHU_V2:

 Say Y here if you want to build the ARM MHUv2 controller driver,
 which provides unidirectional mailboxes between processing elements.

 Symbol: ARM_MHU_V2 [=n]
 Type  : tristate
 Defined at drivers/mailbox/Kconfig:19
   Prompt: ARM MHUv2 Mailbox
   Depends on: MAILBOX [=y] && ARM_AMBA [=y]
   Location:
     -> Device Drivers
       -> Mailbox Hardware Support (MAILBOX [=y])

---

Cc: Al Stone <ahs3@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>