efi: The EFI_VARS is legacy and now x86 only
As of upstream commit 963fabf the EFI_VARS option is set to x86
only as it's long been deprecated pre !intel support and is due to
be removed so it now does nothing on arm platforms.

Fedora actually disabled this on x86 too, I do wonder if it's
worth considering disabling it in ark platforms too?

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>
Cc: Mark Salter <msalter@redhat.com>
Cc: Al Stone <ahs3@redhat.com>