[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_UV_SYSFS:

 This driver supports a sysfs tree describing information about
 UV systems at /sys/firmware/sgi_uv/.

 To compile this driver as a module, choose M here: the module will
 be called uv_sysfs.

 Symbol: UV_SYSFS [=n]
 Type  : tristate
 Defined at drivers/platform/x86/Kconfig:81
   Prompt: Sysfs structure for UV systems
   Depends on: X86 [=y] && X86_PLATFORM_DEVICES [=y] && X86_UV [=y] && SYSFS [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])

---

Cc: David Arcari <darcari@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>