[redhat] New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ZCRYPT_DEBUG:

 Say 'Y' here to enable some additional debug features on the
 s390 cryptographic adapters driver.

 There will be some more sysfs attributes displayed for ap cards
 and queues and some flags on crypto requests are interpreted as
 debugging messages to force error injection.

 Do not enable on production level kernel build.

 If unsure, say N.

 Symbol: ZCRYPT_DEBUG [=n]
 Type  : bool
 Defined at drivers/crypto/Kconfig:78
   Prompt: Enable debug features for s390 cryptographic adapters
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && DEBUG_KERNEL [=y] && ZCRYPT [=m]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
         -> Support for s390 cryptographic adapters (ZCRYPT [=m])

---

Cc: Herbert Xu <herbert.xu@redhat.com>
Cc: "David S. Miller" <davem@redhat.com>
Cc: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>