[redhat] Makefile: new dist-buildreq-check target
Add a new dist-buildreq-check target that checks for missing
packages via rpmspec -q --buildrequires, and make dist-configs-prep
dependent on it.

This change is based on suggestions from Josh Poimboeuf, Don Zickus,
and Herton Krzesinski.  Ultimately this is Herton's version, which
generates the list of missing packages from an already-created
kernel.spec file.

Signed-off-by: Ben Crocker <bcrocker@redhat.com>