New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_KVM_ARM_PMU:

 Adds support for a virtual Performance Monitoring Unit (PMU) in
 virtual machines.

 Symbol: KVM_ARM_PMU [=y]
 Type  : bool
 Defined at arch/arm64/kvm/Kconfig:52
   Prompt: Virtual Performance Monitoring Unit (PMU) support
   Depends on: VIRTUALIZATION [=y] && KVM [=y] && HW_PERF_EVENTS [=y]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Kernel-based Virtual Machine (KVM) support (KVM [=y])

Cc: Mark Salter <msalter@redhat.com>