redhat/configs: enable IMA_KEXEC for supported arches
In RHEL-8, powerpc has support for IMA_KEXEC, and today, only ppc and
aarch64 supports IMA_KEXEC, which are backed by the HAVE_IMA_KEXEC option.
With that, lets enable it in ARK and also, since the HAVE_* option it's safe
to enable the CONFIG_IMA_KEXEC option as a common/generic option too.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>