Add --with clang_lto option to build the kernel with Link Time Optimizations
This also requires the --with toolchain_clang option in order to be
fully enabled.  Currently only supported with x86_64 and aarch64.

As a part of this change, the process_configs.sh script has been updated
so that it can be passed additional make options via the -m flag.  The
LTO related config options depend on CONFIG_CC_IS_CLANG, and so the make
invocations in this script need to be passed CC=clang in order for the LTO
options to be enabled.

Signed-off-by: Tom Stellard <tstellar@redhat.com>