redhat/configs: Enable ARCH_TEGRA on RHEL
ARCH_TEGRA and SERIAL_8250_TEGRA are enabled in RHEL8.4 and should
be enabled in ark as well.

Fedora currently enables these so I moved those into common/generic/arm.
Enabling ARCH_TEGRA uncovered a number of other Tegra options which needed
to be added for ark (none of which enabled at this point).

Signed-off-by: Mark Salter <msalter@redhat.com>