redhat: configs: Enable CONFIG_THINKPAD_LMI
If the Lenovo BIOS lock-bootorder option is enabled then fwupd
cannot install EFI capsule updates as those require adding
temporary boot-option pointing to the update.

The latest fwupd code has support for detecting this through the
think-lmi kernel module, so this should be enabled for full
fwupd functionality.

And this is also something which I expect Lenovo to want to have
enabled in RHEL kernels in general.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v2:
-Move CONFIG_THINKPAD_LMI to configs/common/generic/x86