CI: use more templates
Since the initial implementation, we've added a few more common options
and anchors into the kernel templates. Relevant here is .rhel_common
which we can now use, so we don't forget to update any of the options.

Also split the lines of the "extends" lists to make a visual distinction
between the common (imported) anchors and the ones defined in this CI
file.

Bugzilla: INTERNAL
Upstream Status: RHEL-only

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>