[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SENSORS_MAX6620 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:1035
   Prompt: Maxim MAX6620 fan controller
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])

---

Cc: Dean Nelson <dnelson@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>