[redhat] New configs in drivers/net/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMT [=n]
 Type  : tristate
 Defined at drivers/net/Kconfig:294
   Prompt: Automatic Multicast Tunneling (AMT)
   Depends on: NETDEVICES [=y] && NET_CORE [=y] && INET [=y] && IP_MULTICAST [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Network core driver support (NET_CORE [=y])
 Selects: NET_UDP_TUNNEL [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>