[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GPIO_SIM [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:1684
   Prompt: GPIO Simulator Module
   Depends on: GPIOLIB [=y]
   Location:
     Main menu
       -> Device Drivers
         -> GPIO Support (GPIOLIB [=y])
           -> Virtual GPIO drivers
 Selects: IRQ_SIM [=n] && CONFIGFS_FS [=y]

---

Cc: Brian Masney <bmasney@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>