spec: speed up "cp -r" when it overwrites existing files.
On "slow" storage, this reduces "cp -r" time from
5-15 minutes to a few seconds.

>> and if file exists, open(O_WRONLY|O_TRUNC).
>
>Great. cp uses a common unsafe overwrite anti-pattern. i.e. it can
>result in an empty file (losing the old data without having any of
>the new data written) if the system crashes at just the wrong time.
>
>As it is, open(O_WRONLY|O_TRUNC) on an existing file of 177 bytes in
>length is a truncate down (same as open(O_WRONLY); ftruncate(0);),
>so that would definitely cause XFS to trigger the XFS_ITRUNCATED
>flush-on-close() data loss mitigation here.

Why won't we just use mv instead of cp, you ask?
(1) There is no "mv -r", you need to open-code the logic
to recurse the directories, and move only files.
And more importantly (2) renames on these filesystems also
wait syncronously for metadata update to hit the storage.
I tested it:

    (cd restore && find -type d) \
    | while read -r dir; do
        mkdir -p "lib/modules/$KernelVer/$dir"
        find "restore/$dir" -maxdepth 1 -type f -exec mv -t "lib/modules/$KernelVer/$dir" '{}' '+'
    done

It is about as slow as the "cp -r".

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>