Print arch with process_configs errors
Now that we are capable of processing multiple config files in parallel,
we do not know exactly which arch a failure is seen on. This is
important as frequently errors are specific to a single arch. This patch
just prints the arch before the variant on failure messages to help us
determine the best way to correct an issue.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>