[redhat] New configs in drivers/thermal
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_HFI_THERMAL [=n]
 Type  : bool
 Defined at drivers/thermal/intel/Kconfig:103
   Prompt: Intel Hardware Feedback Interface
   Depends on: THERMAL [=y] && (X86 [=y] || X86_INTEL_QUARK [=n] || COMPILE_TEST [=n]) && NET [=y] && CPU_SUP_INTEL [=y] && X86_THERMAL_VECTOR [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Thermal drivers (THERMAL [=y])
           -> Intel thermal drivers
 Selects: THERMAL_NETLINK [=y]

---

Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_INTEL_HFI_THERMAL enabled under ark/generic/x86/.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>