[redhat] New configs in security/keys
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: USER_DECRYPTED_DATA [=n]
 Type  : bool
 Defined at security/keys/Kconfig:109
   Prompt: Allow encrypted keys with user decrypted data
   Depends on: ENCRYPTED_KEYS [=y]
   Location:
     Main menu
       -> Security options
         -> Enable access key retention support (KEYS [=y])
           -> ENCRYPTED KEYS (ENCRYPTED_KEYS [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>