[redhat] New configs in drivers/mtd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MTD_NAND_ECC_MXIC [=n]
 Type  : bool
 Defined at drivers/mtd/nand/Kconfig:49
   Prompt: Macronix external hardware ECC engine
   Depends on: MTD [=m] && HAS_IOMEM [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Memory Technology Device (MTD) support (MTD [=m])
           -> NAND
             -> ECC engine support
 Selects: MTD_NAND_ECC [=n]
 Implied by [n]:
   - SPI_MXIC [=n] && SPI [=y] && SPI_MASTER [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>