[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_ERRATUM_1902691 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:853
   Prompt: Cortex-A510: 1902691: workaround TRBE trace corruption
   Depends on: CORESIGHT_TRBE [=m]
   Location:
     Main menu
       -> Kernel Features
         -> ARM errata workarounds via the alternatives framework

---

 Symbol: ARM64_ERRATUM_2038923 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:830
   Prompt: Cortex-A510: 2038923: workaround TRBE corruption with enable
   Depends on: CORESIGHT_TRBE [=m]
   Location:
     Main menu
       -> Kernel Features
         -> ARM errata workarounds via the alternatives framework

---

 Symbol: ARM64_ERRATUM_2064142 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:812
   Prompt: Cortex-A510: 2064142: workaround TRBE register writes while disabled
   Depends on: CORESIGHT_TRBE [=m]
   Location:
     Main menu
       -> Kernel Features
         -> ARM errata workarounds via the alternatives framework

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>