Redhat: enable Kfence on production servers
Kfence allows for a low overhead memory error detection system that can
be deployed in production. By enabling this feature we allow for better
and quicker bug reporting/fixing.

the Kscale team has done some performance testing and have concluded
there is no noticable performance impact.

Enable for ELN and remove the Fedora specific config definition.
We may want to consider enabling CONFIG_KFENCE_DEFERRABLE on fedora...
This config allows the CPU wakup to be deferred, which is more ideal for
power-constrianted systems which is more likely to be the case on Fedora.

Signed-off-by: Nico Pache npache@redhat.com