[redhat] New configs in crypto/asymmetric_keys
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FIPS_SIGNATURE_SELFTEST [=n]
 Type  : bool
 Defined at crypto/asymmetric_keys/Kconfig:78
   Prompt: Run FIPS selftests on the X.509+PKCS7 signature verification
   Depends on: CRYPTO [=y] && KEYS [=y] && ASYMMETRIC_KEY_TYPE [=y] && PKCS7_MESSAGE_PARSER [=y]
   Location:
     Main menu
       -> Cryptographic API (CRYPTO [=y])
         -> Asymmetric (public-key cryptographic) key type (ASYMMETRIC_KEY_TYPE [=y])

---

Cc: Herbert Xu <herbert.xu@redhat.com>
Cc: "David S. Miller" <davem@redhat.com>
Cc: Vladis Dronov <vdronov@redhat.com>
Cc: Bruno Meneguele <bmeneg@redhat.com>
Cc: Ondrej Mosnacek <omosnace@redhat.com>
Cc: Phil Auld <pauld@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>