[redhat] New configs in drivers/pwm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PWM_CLK [=n]
 Type  : tristate
 Defined at drivers/pwm/Kconfig:143
   Prompt: Clock based PWM support
   Depends on: PWM [=y] && (HAVE_CLK [=y] || COMPILE_TEST [=n])
   Location:
     Main menu
       -> Device Drivers
         -> Pulse-Width Modulation (PWM) Support (PWM [=y])

---

Cc: Doug Ledford <dledford@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>