Leave CONFIG_ACPI_VIDEO on for x86 only
Upstream commit 038275d227841 has removed the dependency on X86 for
ACPI_VIDEO. This means the config has to be set for all non x86 arches
as well. This change will turn off ACPI_VIDEO for all but x86 leaving us
with similar behavior to what we had in 5.19 and prior.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>