[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TYPEC_ANX7411 [=n]
 Type  : tristate
 Defined at drivers/usb/typec/Kconfig:55
   Prompt: Analogix ANX7411 Type-C DRP Port controller driver
   Depends on: USB_SUPPORT [=y] && TYPEC [=y] && I2C [=y] && USB_ROLE_SWITCH [=y]
   Location:
     Main menu
       -> Device Drivers
         -> USB support (USB_SUPPORT [=y])
           -> USB Type-C Support (TYPEC [=y])

---

 Symbol: UCSI_STM32G0 [=n]
 Type  : tristate
 Defined at drivers/usb/typec/ucsi/Kconfig:51
   Prompt: UCSI Interface Driver for STM32G0
   Depends on: USB_SUPPORT [=y] && TYPEC [=y] && TYPEC_UCSI [=y] && I2C [=y]
   Location:
     Main menu
       -> Device Drivers
         -> USB support (USB_SUPPORT [=y])
           -> USB Type-C Support (TYPEC [=y])
             -> USB Type-C Connector System Software Interface driver (TYPEC_UCSI [=y])

---

 Symbol: USB_ONBOARD_HUB [=n]
 Type  : tristate
 Defined at drivers/usb/misc/Kconfig:299
   Prompt: Onboard USB hub support
   Depends on: USB_SUPPORT [=y] && USB [=y] && (OF [=y] || COMPILE_TEST [=n])
   Location:
     Main menu
       -> Device Drivers
         -> USB support (USB_SUPPORT [=y])

---

Cc: Torez Smith <torez@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>