[redhat] New configs in drivers/reset
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RESET_TI_TPS380X [=n]
 Type  : tristate
 Defined at drivers/reset/Kconfig:269
   Prompt: TI TPS380x Reset Driver
   Depends on: RESET_CONTROLLER [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Reset Controller Support (RESET_CONTROLLER [=y])
 Selects: GPIOLIB [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>