redhat: always add a rebase entry when ark merges from upstream
We have some complex logic to parse changelog entries to choose whether
we add a rebase entry or not to the changelog. Since genlog.py is adding
a Resolves entry always now, simply replace that logic with a
replacement using sed to add a rebase entry when the markers differ.
The marker check doesn't work at the moment due Makefile logic but
that will be fixed in the following changes.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>