redhat: fix release number bump when dist-release-changed runs
Right now inside the dist-release-changed target, we run a sub make
invocation calling the dist-release-finish target, after bumping
the release number with new_release.sh. The problem is that the
BUILD number still pointed to the old pre-bump number inside the
dist-release-finish invocation, and the problem is that we don't
reload the RHEL_RELEASE number in redhat/Makefile if BUILD is already
defined. Since it looks like that check was added to avoid overriding
the BUILD number set when RHSELFTESTDATA is defined, lets change
the if check there so BUILD has the updated value when dist-release-finish
runs through dist-release-changed.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>