redhat/kernel.spec.template: enable dependencies generation
Updated kernel-rpm-macros-185-9 package brings various generators
for Provides tags that usually present in RHEL kernel packages,
including kernel(), which OOT kmod packages rely on.

prarit: After discussion with jforbes, it was found that kernel-rpm-macros
has a different version number in Fedora vs RHEL.  Additionally,
kernel-rpm-macros is brought in through some other build dependency in
ARK/Fedora so the requirement here may be superfluous.  As a result, I am
only mentioning the build requirement of the package in this commit and
not the version.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>