Common config cleanup for 6.0
The is the output of evaluate_configs -j which compares config entries
in fedora and ark directories, consolidating them to common if they are
identical. It has been verified that there is 0 diff in the generated
kernel-*.config files with this change. There is an unusually large
number of clean-ups in the x86 directories because those directories
were also cleaned up with a previous MR which got rid of the x86_64
directory.  As a result, an entry which was set in x86 in Fedora and in
x86_64 in ark are now both in the same x86 directory, and are evaluated
for a match.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>