[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INPUT_IBM_PANEL [=n]
 Type  : tristate
 Defined at drivers/input/misc/Kconfig:733
   Prompt: IBM Operation Panel driver
   Depends on: INPUT [=y] && INPUT_MISC [=y] && I2C [=y] && I2C_SLAVE [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Input device support
           -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
             -> Miscellaneous devices (INPUT_MISC [=y])

---

 Symbol: TOUCHSCREEN_COLIBRI_VF50 [=n]
 Type  : tristate
 Defined at drivers/input/touchscreen/Kconfig:1336
   Prompt: Toradex Colibri on board touchscreen driver
   Depends on: INPUT [=y] && INPUT_TOUCHSCREEN [=y] && IIO [=m] && (GPIOLIB [=y] || COMPILE_TEST [=n])
   Location:
     Main menu
       -> Device Drivers
         -> Input device support
           -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
             -> Touchscreens (INPUT_TOUCHSCREEN [=y])

---

Cc: Tony Camuso <tcamuso@redhat.com>
Cc: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>