redhat/Makefile: Rename LOCALVERSION to DISTLOCALVERSION
LOCALVERSION is actually a name already used in the kernel for
the last version suffix. Not only it is confusing for programmers,
but it also confuses scripts/setlocalversion and changes the
kernel version, causing mismatches with e.g. the name of the
directory where modules are installed and resulting in module
load failures.

Rename it to DISTLOCALVERSION to avoid the name collision.

Upstream Status: RHEL only
Suggested-by: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>