configs/fedora: Build some SC7180 clock controllers as modules
The clock controllers for the Camera, Display, GPU and Video {en,de}coder
don't need to be built-in and can be built as a module instead. Do that,
which is also aligned with how the clock controller drivers are enabled
for the SC7280 SoC.

Also, cleanup the clk Kconfig symbols not set in fedora/generic/ and move
all the SC7180 clock controller related symbols to fedora/generic/arm dir.

Suggested-by: Brian Masney <bmasney@redhat.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>