CI: Remove old configs
We're not using this file for c9s/RHEL9 configuration anymore and can
thus drop all related configs. It's not worth it to keep them around for
c10s forking, as they are out of fate compared to the regularly updated
configs in the c9s repo. When the forking time comes, the updated
configs from there should be used as a base for c10s instead of these
old ones.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>