[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KALLSYMS_SELFTEST [=n]
 Type  : bool
 Defined at init/Kconfig:1726
   Prompt: Test the basic functions and performance of kallsyms
   Depends on: KALLSYMS [=y]
   Location:
     -> General setup
       -> Load all symbols for debugging/ksymoops (KALLSYMS [=y])
         -> Test the basic functions and performance of kallsyms (KALLSYMS_SELFTEST [=n])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Artem Savkov <asavkov@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Jan Stancek <jstancek@redhat.com>
Cc: Josh Poimboeuf <jpoimboe@kernel.org>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>