[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTERRUPT_SANITIZE_REGISTERS [=y]
 Type  : bool
 Defined at arch/powerpc/Kconfig:539
   Prompt: Clear gprs on interrupt arrival
   Depends on: PPC64 [=y] && ARCH_HAS_SYSCALL_WRAPPER [=y]
   Location:
     -> Kernel options
       -> Clear gprs on interrupt arrival (INTERRUPT_SANITIZE_REGISTERS [=y])

---

Cc: Steve Best <sbest@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>