[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_SMPRO [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:80
   Prompt: Ampere Computing SMpro core driver
   Depends on: HAS_IOMEM [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Ampere Computing SMpro core driver (MFD_SMPRO [=n])
 Selects: MFD_CORE [=y] && REGMAP_I2C [=m]

---

 Symbol: MFD_TPS65219 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:1628
   Prompt: TI TPS65219 Power Management IC
   Depends on: HAS_IOMEM [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> TI TPS65219 Power Management IC (MFD_TPS65219 [=n])
 Selects: MFD_CORE [=m] && REGMAP_I2C [=m] && REGMAP_IRQ [=n]

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>