Turn on CONFIG_XFS_RT for Fedora
While this was initially turned off as it came in as experimental, it
has not been labelled as such for quite a while now. We had a requrest
to turn it on.  As it was set to off in common, the file was moved from
common to RHEL to ensure that RHEL configs did not change, and a new
file was added to Fedora turning it on.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>