redhat/kernel.spec.template: fix installonlypkg for meta package
The meta package is missing 'installonlypkg' Provides. This causes an
issue with upgrade path in downstream RHEL for 'kernel-rt' (RHBZ 2177242):
  cannot install both kernel-rt-5.14.0-286.el9.x86_64 and kernel-rt-5.14.0-284.1.1.rt14.285.el9_2.x86_64

The cause was identified as mix of packages, where some have
installonlypkg Provides and some don't.

The Provides is missing in Fedora meta package too:
  # rpm -qp --provides ./kernel-6.3.0-0.rc2.20230315git6015b1aca1a2.25.fc39.x86_64.rpm
  kernel = 6.3.0-0.rc2.20230315git6015b1aca1a2.25.fc39
  kernel(x86-64) = 6.3.0-0.rc2.20230315git6015b1aca1a2.25.fc39

Tested: With RHEL 9.2 kernel-rt build patched and installed, install of
        new 9.3 kernel-rt works.

Signed-off-by: Jan Stancek <jstancek@redhat.com>