[redhat] New configs in drivers/base
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: REGMAP_KUNIT [=m]
 Type  : tristate
 Defined at drivers/base/regmap/Kconfig:12
   Prompt: KUnit tests for regmap
   Depends on: KUNIT [=m]
   Location:
     -> Device Drivers
       -> Generic Driver Options
         -> KUnit tests for regmap (REGMAP_KUNIT [=m])
 Selects: REGMAP [=y] && REGMAP_RAM [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>