[redhat] New configs in drivers/net/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MICROCHIP_T1S_PHY [=n]
 Type  : tristate
 Defined at drivers/net/phy/Kconfig:245
   Prompt: Microchip 10BASE-T1S Ethernet PHY
   Depends on: NETDEVICES [=y] && PHYLIB [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> PHY Device support and infrastructure (PHYLIB [=y])
           -> Microchip 10BASE-T1S Ethernet PHY (MICROCHIP_T1S_PHY [=n])

---

 Symbol: NXP_CBTX_PHY [=n]
 Type  : tristate
 Defined at drivers/net/phy/Kconfig:279
   Prompt: NXP 100BASE-TX PHYs
   Depends on: NETDEVICES [=y] && PHYLIB [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> PHY Device support and infrastructure (PHYLIB [=y])
           -> NXP 100BASE-TX PHYs (NXP_CBTX_PHY [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>