[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: COMMON_CLK_SI521XX [=n]
 Type  : tristate
 Defined at drivers/clk/Kconfig:379
   Prompt: Clock driver for SkyWorks Si521xx PCIe clock generators
   Depends on: COMMON_CLK [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])
         -> Clock driver for SkyWorks Si521xx PCIe clock generators (COMMON_CLK_SI521XX [=n])
 Selects: REGMAP_I2C [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>