redhat: configs: generic: x86: Disable CONFIG_VIDEO_OV2740 for x86 platform
For the Intel MIPI camera solution, OV2740 sensor is used to provide
video streaming to IPU through the MIPI interface. However, the upstream
already has the sensor's driver to drive the device running. The upstream
driver can't be used for this MIPI camera solution and it also causes
several errors. Intel also provides a opensource driver which integrates
OV2740 with Intel MIPI camera and INT3472. To reduce the configuration
overhead for the users, the OV2740 could be disabled for the X86 platform.
This driver will be installed through the IPU6 camera akmod package.

Signed-off-by: Kate Hsuan <hpa@redhat.com>