redhat/kernel.spec.template: Disable 'extracting debug info' messages
debugedit's find_debuginfo.sh script outputs an 'extracting debug info'
message for every binary file that is packaged.  On a recent ELN build
these messages are responsible for 4631 out of 11464 (40%) of build log
messages on x86_64 [1].  They serve little or no purpose and are noise in the
build log.  These messages have already been disabled in Fedora, so
disable them for CentOS and ELN builds as well.

Disable 'extracting debug info' messages for all builds and add a
note on how to re-enable them if necessary.

[1] https://koji.fedoraproject.org/koji/buildinfo?buildID=2218314

Signed-off-by: Prarit Bhargava <prarit@redhat.com>