config: wifi: debug options for ath11k, brcm80211 and iwlwifi
Set CONFIGs to enable the debug options for ath11k and brcm80211, both
for Fedora and RHEL.

Enable iwlwifi debug options only for kernel-debug in RHEL, and disable them
for the standard kernel, so they are in line with the other drivers.
Leave them unthouched in Fedora, despite being enabled for the normal
kernel, because there might be any reason I am not aware of.

Also, cleanup some iwlwifi debug configs, moving to common those that
are common to Fedora and RHEL.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>