[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_AMDGPU_WERROR [=n]
 Type  : bool
 Defined at drivers/gpu/drm/amd/amdgpu/Kconfig:72
   Prompt: Force the compiler to throw an error instead of a warning when compiling
   Depends on: HAS_IOMEM [=y] && DRM_AMDGPU [=m] && EXPERT [=y] && !COMPILE_TEST [=n]
   Location:
     -> Device Drivers
       -> Graphics support
         -> AMD GPU (DRM_AMDGPU [=m])
           -> Force the compiler to throw an error instead of a warning when compiling (DRM_AMDGPU_WERROR [=n])

---

 Symbol: DRM_PANEL_SAMSUNG_S6D7AA0 [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/panel/Kconfig:556
   Prompt: Samsung S6D7AA0 MIPI-DSI video mode panel controller
   Depends on: HAS_IOMEM [=y] && DRM [=y] && DRM_PANEL [=y] && OF [=y] && BACKLIGHT_CLASS_DEVICE [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Panels
           -> Samsung S6D7AA0 MIPI-DSI video mode panel controller (DRM_PANEL_SAMSUNG_S6D7AA0 [=n])
 Selects: DRM_MIPI_DSI [=n] && VIDEOMODE_HELPERS [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>