[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_JITTERENTROPY_TESTINTERFACE [=n]
 Type  : bool
 Defined at crypto/Kconfig:1299
   Prompt: CPU Jitter RNG Test Interface
   Depends on: CRYPTO [=y] && CRYPTO_JITTERENTROPY [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Random number generation
         -> CPU Jitter Non-Deterministic RNG (Random Number Generator) (CRYPTO_JITTERENTROPY [=y])
           -> CPU Jitter RNG Test Interface (CRYPTO_JITTERENTROPY_TESTINTERFACE [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>