[redhat] New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_DEV_FSL_CAAM_RNG_TEST [=n]
 Type  : bool
 Defined at drivers/crypto/caam/Kconfig:165
   Prompt: Test caam rng
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && CRYPTO_DEV_FSL_CAAM [=m] && CRYPTO_DEV_FSL_CAAM_JR [=m]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
         -> Freescale CAAM-Multicore platform driver backend (CRYPTO_DEV_FSL_CAAM [=m])
           -> Freescale CAAM Job Ring driver backend (CRYPTO_DEV_FSL_CAAM_JR [=m])
             -> Test caam rng (CRYPTO_DEV_FSL_CAAM_RNG_TEST [=n])
 Selects: CRYPTO_DEV_FSL_CAAM_RNG_API [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>