[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_QCOM_SGMII_ETH [=n]
 Type  : tristate
 Defined at drivers/phy/qualcomm/Kconfig:192
   Prompt: Qualcomm DWMAC SGMII SerDes/PHY driver
   Depends on: OF [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> Qualcomm DWMAC SGMII SerDes/PHY driver (PHY_QCOM_SGMII_ETH [=n])
 Selects: GENERIC_PHY [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>