redhat/scripts/expand_srpm.sh: add missing function, robustify
"die" (presumably a shell function) is not defined. Add it.

Delete the empty line at the end of the script,
git complained when I added a copy of this file to another tree.

This part will happily execute if someone just accidentally runs
the script from the command line with no arguments:
    cd "$cloned/$PACKAGE_NAME"
    rm -r -- *
...and might delete most of your files!
Add a safeguard against this.

While at it, remove superfluous stray semicolons.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>