fedora: remove ARMv7 AKA armhfp support
The ARMv7 architecture was retired in Fedora 37 so now with
the Fedora 36 release End of Life we now no longer need the
kernel pieces in place so let's do an initial cleanup. This
removes the build process pieces. This also removes the last
vestiges of PAE kernels as ARMv7 was the last user of them.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>