Apply partial snippet configs to all configs
There is a macro all_arch_configs that tries to optimize what configs we
apply the partial snippets to.  While it speeds up things up a bit, it
creates conflicts with CLANG options.  The process_configs.sh script
still processes _all_ the arches and expects the CLANG options to be set
correctly.

I can either add in an arch specific option to process_configs.sh to
only process the configs the correct CLANG snippet was applied to _or_
just apply the snippets to _all_ the configs.

The latter is easier so that is what this change does.  The cost is an
extra second to apply the snippets to more config files.

Signed-off-by: Don Zickus <dzickus@redhat.com>