[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GPIO_DS4520 [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:1038
   Prompt: DS4520 I2C GPIO expander
   Depends on: GPIOLIB [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> I2C GPIO expanders
           -> DS4520 I2C GPIO expander (GPIO_DS4520 [=n])
 Selects: REGMAP_I2C [=m] && GPIO_REGMAP [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>