[redhat] New configs in drivers/watchdog
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MARVELL_GTI_WDT [=y]
 Type  : tristate
 Defined at drivers/watchdog/Kconfig:1803
   Prompt: Marvell GTI Watchdog driver
   Depends on: WATCHDOG [=y] && (ARCH_THUNDER [=y] || COMPILE_TEST [=n] && 64BIT [=y])
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])
         -> Marvell GTI Watchdog driver (MARVELL_GTI_WDT [=y])
 Selects: WATCHDOG_CORE [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>