[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KVM_PROVE_MMU [=n]
 Type  : bool
 Defined at arch/x86/kvm/Kconfig:141
   Prompt: Prove KVM MMU correctness
   Depends on: VIRTUALIZATION [=y] && DEBUG_KERNEL [=y] && KVM [=m] && EXPERT [=y]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Kernel-based Virtual Machine (KVM) support (KVM [=m])
         -> Prove KVM MMU correctness (KVM_PROVE_MMU [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>