[redhat] New configs in lib/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LWQ_TEST [=n]
 Type  : bool
 Defined at lib/Kconfig:735
   Prompt: Boot-time test for lwq queuing
   Location:
     -> Library routines
       -> Boot-time test for lwq queuing (LWQ_TEST [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>