[redhat] New configs in net/ipv4
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TCP_AO [=n]
 Type  : bool
 Defined at net/ipv4/Kconfig:747
   Prompt: TCP: Authentication Option (RFC5925)
   Depends on: NET [=y] && INET [=y] && 64BIT [=y] && IPV6 [=y]!=m
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> TCP/IP networking (INET [=y])
           -> TCP: Authentication Option (RFC5925) (TCP_AO [=n])
 Selects: CRYPTO [=y] && TCP_SIGPOOL [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_TCP_AO enabled

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>