redhat: disable dccp conntrack support
DCCP is already disabled in RHEL.

Judging from activity on the upstream development mailing list and the
"syszkaller" dccp related reports there are no real users of this protocol.

Disable dccp conntrack support.  We've had use-after-free and out-of-bounds
acesses (now fixed) in this code in the past, its clear that this is
only "used" by fuzzers.

Signed-off-by: Florian Westphal <fwestpha@redhat.com>