[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PCIEAER_CXL [=y]
 Type  : bool
 Defined at drivers/pci/pcie/Kconfig:52
   Prompt: PCI Express CXL RAS support
   Depends on: PCI [=y] && PCIEAER [=y] && CXL_PCI [=m]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI Express Port Bus support (PCIEPORTBUS [=y])
           -> PCI Express Advanced Error Reporting support (PCIEAER [=y])
             -> PCI Express CXL RAS support (PCIEAER_CXL [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>