common: disable the FB device creation
With the move to Simple DRM and associated SYSFB we no
longer need the /dev/fb* devices even for fallback
scenarios or consoles on any supported architectures
so lets disable them as there's a security implication
as outlined in kernel commit 701d2054.

This does not affect FB console emulation as done on
top of the DRM stack for kernel and related output.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>