[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TRANSPARENT_HUGEPAGE_NEVER [=n]
 Type  : bool
 Defined at mm/Kconfig:839
   Prompt: never
   Depends on: <choice>
   Location:
     -> Memory Management options
       -> Transparent Hugepage Support (TRANSPARENT_HUGEPAGE [=y])
         -> Transparent Hugepage Support sysfs defaults (<choice> [=y])
           -> never (TRANSPARENT_HUGEPAGE_NEVER [=n])

---

 Symbol: ZSWAP_SHRINKER_DEFAULT_ON [=n]
 Type  : bool
 Defined at mm/Kconfig:64
   Prompt: Shrink the zswap pool on memory pressure
   Depends on: ZSWAP [=y]
   Location:
     -> Memory Management options
       -> Support for paging of anonymous memory (swap) (SWAP [=y])
         -> Compressed cache for swap pages (ZSWAP [=y])
           -> Shrink the zswap pool on memory pressure (ZSWAP_SHRINKER_DEFAULT_ON [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>