Remove CONFIG_NET_EMATCH_STACK file for RHEL
This file was set to:
CONFIG_NET_EMATCH_STACK is not set

Unofortunately that is not a valid entry for this. It would require an
integer.  In the past, Kconfig didn't care so much, but recent changes
make Kconfig error out on this and kill the build. We can simply delete
the entry because CONFIG_NET_EMATCH is not set for RHEL either, and that
is the only thing with a dep on CONFIG_NET_EMATCH_STACK.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>