redhat/Makefile: fix setup-source and document its caveat
The current definition of setup-source is racy, since it may cause
dist-clean-sources and _setup-source to be "built" in parallel and the
former caould then remove the results of the latter. Fix it to ensure
the right ordering and document the condition that must be preserved to
keep dist-clean-sources from racing with other targets.

_setup-source (and this race) has been introduced in commit 0cce4209d1cc
("redhat/Makefile: Split up setup-source target") in order to be called
from redhat/self-test/data/create-data.sh, but that script can
apparently work with the cleanup as well, so we can remove it and
replace it with setup-source.

Fixes: 0cce4209d1cc ("redhat/Makefile: Split up setup-source target")
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>