redhat/configs: For aarch64/RT, default kstack randomization off
Unlike other architectures which use their cycles counters to
seed the kstack randomization (which is sketchy) aarch64 uses
the kernel random number generator. For a list of reasons starting
with the lack of an enabled/fast cycle counter, a ubiquitous high
frequency timer, nor machines which have the v8.5 FEAT_RNG.

Using the kernel rng is largely an acceptable solution except
for RT workloads which can experience quite large latency spikes
as a result.

So, lets invert the default kstack randomization for RT kernels
on aarch64 until FEAT_RNG is common or a better solution is
found.

Signed-off-by: Jeremy Linton <jlinton@redhat.com>