Consolidate 6.8 configs to common
As is done most releases, this is the output of evaluate_configs -j.
It has been verified that the generated config files are identical to
config files generated just before the script was run. If you are
looking for history on a file added to common in this commit, it would
have existed in both the rhel and fedora directories before.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>