[redhat] New configs in drivers/ras
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMD_ATL [=m]
 Type  : tristate
 Defined at drivers/ras/amd/atl/Kconfig:10
   Prompt: AMD Address Translation Library
   Depends on: AMD_NB [=y] && X86_64 [=y] && RAS [=y] && MEMORY_FAILURE [=y]
   Location:
     -> Device Drivers
       -> Reliability, Availability and Serviceability (RAS) features (RAS [=y])
         -> AMD Address Translation Library (AMD_ATL [=m])
 Implied by [m]:
   - EDAC_AMD64 [=m] && EDAC [=y] && AMD_NB [=y] && EDAC_DECODE_MCE [=m]

 Commit: 3f3174996be6 (RAS: Introduce AMD Address Translation Library)

---

 Symbol: RAS_FMPM [=m]
 Type  : tristate
 Defined at drivers/ras/Kconfig:37
   Prompt: FRU Memory Poison Manager
   Depends on: RAS [=y] && AMD_ATL [=m] && ACPI_APEI [=y]
   Location:
     -> Device Drivers
       -> Reliability, Availability and Serviceability (RAS) features (RAS [=y])
         -> AMD Address Translation Library (AMD_ATL [=m])
           -> FRU Memory Poison Manager (RAS_FMPM [=m])

 Commit: 6f15e617cc99 (RAS: Introduce a FRU memory poison manager)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>