[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MITIGATION_RFDS [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:2627
   Prompt: RFDS Mitigation
   Depends on: SPECULATION_MITIGATIONS [=y] && CPU_SUP_INTEL [=y]
   Location:
     -> Mitigations for speculative execution vulnerabilities (SPECULATION_MITIGATIONS [=y])
       -> RFDS Mitigation (MITIGATION_RFDS [=y])

 Commit: 8076fcde016c (x86/rfds: Mitigate Register File Data Sampling (RFDS))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>