[redhat] New configs in drivers/net/can
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CAN_ESD_402_PCI [=n]
 Type  : tristate
 Defined at drivers/net/can/esd/Kconfig:2
   Prompt: esd electronics gmbh CAN-PCI(e)/402 family
   Depends on: NETDEVICES [=y] && CAN_DEV [=m] && CAN_NETLINK [=y] && PCI [=y] && HAS_DMA [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> CAN Device Drivers (CAN_DEV [=m])
           -> CAN device drivers with Netlink support (CAN_NETLINK [=y])
             -> esd electronics gmbh CAN-PCI(e)/402 family (CAN_ESD_402_PCI [=n])

 Commit: 9721866f07e1 (can: esd: add support for esd GmbH PCIe/402 CAN interface family)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>