[redhat] New configs in net/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BQL [=y]
 Type  : bool
 Defined at net/Kconfig:332
   Prompt: Enable Byte Queue Limits
   Depends on: NET [=y] && SYSFS [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> Enable Byte Queue Limits (BQL [=y])
 Selects: DQL [=y]

 Commit: 114cf5802165 (bql: Byte queue limits)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>