[redhat] New configs in kernel/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HIBERNATION_COMP_LZO [=y]
 Type  : bool
 Defined at kernel/power/Kconfig:100
   Prompt: lzo
   Depends on: <choice> && CRYPTO_LZO [=y]
   Location:
     -> Power management and ACPI options
       -> Hibernation (aka 'suspend to disk') (HIBERNATION [=y])
         -> Default compressor (<choice> [=y])
           -> lzo (HIBERNATION_COMP_LZO [=y])

 Commit: a06c6f5d3cc9 (PM: hibernate: Move to crypto APIs for LZO compression)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>