[redhat] New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_DEV_QAT_ERROR_INJECTION [=n]
 Type  : bool
 Defined at drivers/crypto/intel/qat/Kconfig:110
   Prompt: Support for Intel(R) QAT Devices Heartbeat Error Injection
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && CRYPTO_DEV_QAT [=m] && DEBUG_FS [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
         -> Support for Intel(R) QAT Devices Heartbeat Error Injection (CRYPTO_DEV_QAT_ERROR_INJECTION [=n])

 Commit: e2b67859ab6e (crypto: qat - add heartbeat error simulator)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>