[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TYPEC_MUX_IT5205 [=n]
 Type  : tristate
 Defined at drivers/usb/typec/mux/Kconfig:39
   Prompt: ITE IT5205 Type-C USB Alt Mode Passive MUX driver
   Depends on: USB_SUPPORT [=y] && TYPEC [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> USB Type-C Support (TYPEC [=y])
           -> USB Type-C Multiplexer/DeMultiplexer Switch support
             -> ITE IT5205 Type-C USB Alt Mode Passive MUX driver (TYPEC_MUX_IT5205 [=n])
 Selects: REGMAP_I2C [=m]

 Commit: 41fe9ea1696c (usb: typec: mux: Add ITE IT5205 Alternate Mode Passive MUX driver)

---

 Symbol: USB_DEFAULT_AUTHORIZATION_MODE [=1]
 Type  : integer
 Range : [0 2]
 Defined at drivers/usb/core/Kconfig:120
   Prompt: Default authorization mode for USB devices
   Depends on: USB_SUPPORT [=y] && USB [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> Default authorization mode for USB devices (USB_DEFAULT_AUTHORIZATION_MODE [=1])

 Commit: bec7e43b162c (usb: core: Make default authorization mode configurable)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>