[redhat] New configs in drivers/tee
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_TSTEE [=n]
 Type  : tristate
 Defined at drivers/tee/tstee/Kconfig:2
   Prompt: Arm Trusted Services TEE driver
   Depends on: TEE [=m] && ARM_FFA_TRANSPORT [=m]
   Location:
     -> Device Drivers
       -> Trusted Execution Environment support (TEE [=m])
         -> Arm Trusted Services TEE driver (ARM_TSTEE [=n])

 Commit: c835e5a3153c (tee: tstee: Add Trusted Services TEE driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>