[redhat] New configs in drivers/bluetooth
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BT_INTEL_PCIE [=n]
 Type  : tristate
 Defined at drivers/bluetooth/Kconfig:481
   Prompt: Intel HCI PCIe driver
   Depends on: NET [=y] && BT [=m] && PCI [=y]
   Location:
     -> Networking support (NET [=y])
       -> Bluetooth subsystem support (BT [=m])
         -> Bluetooth device drivers
           -> Intel HCI PCIe driver (BT_INTEL_PCIE [=n])
 Selects: BT_INTEL [=m] && FW_LOADER [=y]

 Commit: c2b636b3f788 (Bluetooth: btintel_pcie: Add support for PCIe transport)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>