[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: X86_POSTED_MSI [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:469
   Prompt: Enable MSI and MSI-x delivery by posted interrupts
   Depends on: X86_64 [=y] && IRQ_REMAP [=y]
   Location:
     -> Processor type and features
       -> Enable MSI and MSI-x delivery by posted interrupts (X86_POSTED_MSI [=n])

 Commit: 7fec07fd2178 (x86/irq: Add a Kconfig option for posted MSI)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>