[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SOFTLOCKUP_DETECTOR_INTR_STORM [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:1033
   Prompt: Detect Interrupt Storm in Soft Lockups
   Depends on: SOFTLOCKUP_DETECTOR [=y] && IRQ_TIME_ACCOUNTING [=y]
   Location:
     -> Kernel hacking
       -> Debug Oops, Lockups and Hangs
         -> Detect Soft Lockups (SOFTLOCKUP_DETECTOR [=y])
           -> Detect Interrupt Storm in Soft Lockups (SOFTLOCKUP_DETECTOR_INTR_STORM [=n])
 Selects: GENERIC_IRQ_STAT_SNAPSHOT [=n]

 Commit: d7037381d002 (watchdog/softlockup: Low-overhead detection of interrupt storm)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>