New configs in lib/kunit
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KUNIT_FAULT_TEST [=y]
 Type  : bool
 Defined at lib/kunit/Kconfig:27
   Prompt: Enable KUnit tests which print BUG stacktraces
   Depends on: KUNIT [=m] && KUNIT_TEST [=m] && !UML
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> KUnit - Enable support for unit tests (KUNIT [=m])
           -> Enable KUnit tests which print BUG stacktraces (KUNIT_FAULT_TEST [=y])

 Commit: 4b513a02fd05 (kunit: test: Move fault tests behind KUNIT_FAULT_TEST Kconfig option)