[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CLK_IMX95_BLK_CTL [=n]
 Type  : tristate
 Defined at drivers/clk/imx/Kconfig:117
   Prompt: IMX95 Clock Driver for BLK CTL
   Depends on: COMMON_CLK [=y] && (ARCH_MXC [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])
         -> IMX95 Clock Driver for BLK CTL (CLK_IMX95_BLK_CTL [=n])
 Selects: MXC_CLK [=y]

 Commit: 5224b189462f (clk: imx: add i.MX95 BLK CTL clk driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>