[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: USB_ONBOARD_DEV [=n]
 Type  : tristate
 Defined at drivers/usb/misc/Kconfig:319
   Prompt: Onboard USB device support
   Depends on: USB_SUPPORT [=y] && USB [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> Onboard USB device support (USB_ONBOARD_DEV [=n])

 Commit: 31e7f6c015d9 (usb: misc: onboard_hub: rename to onboard_dev)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>