redhat/configs: Re-enable CONFIG_KEXEC for Fedora
Commit c4bf2b6fdfb3 ("Drop kexec_load syscall support") unintentionally
dropped support for the kexec_load system call for both Fedora and RHEL.
However, it can be very useful to have a fallback method when working
close to upstream, e.g. to verify that a bug is in kexec_file_load. Thus
re-enable CONFIG_KEXEC for Fedora but keep it disabled for RHEL.

Signed-off-by: Philipp Rudo <prudo@redhat.com>