[redhat] fix kernel.changelog sorting
We take all the redhat/kernel.changelog-{major}.{minor} files and
generate a unified changelog file from them. It should process the
files in order from newest version to latest version so the unified file
is in reverse chronological order. Unfortunately, the sort operation
orders kernel.changelog-9.99 before kernel.changelog-10.0 so the
resulting file is out of order.

Fix this by using sort's -V (--version-sort) option as it orders the
files as expected.

Fixes: ec7c0d2f5586 ("redhat: ship all the changelog from source git into kernel-doc")

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>