[redhat] New configs in drivers/vdpa
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: OCTEONEP_VDPA [=n]
 Type  : tristate
 Defined at drivers/vdpa/Kconfig:129
   Prompt: vDPA driver for Octeon DPU devices
   Depends on: VDPA [=m] && m [=m] && MODULES [=y] && PCI_MSI [=y]
   Location:
     -> Device Drivers
       -> vDPA drivers (VDPA [=m])
         -> vDPA driver for Octeon DPU devices (OCTEONEP_VDPA [=n])

 Commit: 8b6c724cdab8 (virtio: vdpa: vDPA driver for Marvell OCTEON DPU devices)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>