[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MEMCG_V1 [=n]
 Type  : bool
 Defined at init/Kconfig:974
   Prompt: Legacy cgroup v1 memory controller
   Depends on: CGROUPS [=y] && MEMCG [=y]
   Location:
     -> General setup
       -> Control Group support (CGROUPS [=y])
         -> Memory controller (MEMCG [=y])
           -> Legacy cgroup v1 memory controller (MEMCG_V1 [=n])

 Commit: e93d4166b40a (mm: memcg: put cgroup v1-specific code under a config option)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_MEMCG_V1 enable for RHEL.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>