[redhat] New configs in drivers/irqchip
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LAN966X_OIC [=n]
 Type  : tristate
 Defined at drivers/irqchip/Kconfig:177
   Prompt: Microchip LAN966x OIC Support
   Location:
     -> Device Drivers
       -> IRQ chip support
         -> Microchip LAN966x OIC Support (LAN966X_OIC [=n])
 Selects: GENERIC_IRQ_CHIP [=n] && IRQ_DOMAIN [=y]

 Commit: 3e3a7b353329 (irqchip: Add support for LAN966x OIC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>