[redhat] New configs in drivers/mailbox
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: QCOM_CPUCP_MBOX [=n]
 Type  : tristate
 Defined at drivers/mailbox/Kconfig:279
   Prompt: Qualcomm Technologies, Inc. CPUCP mailbox driver
   Depends on: MAILBOX [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && 64BIT [=y]
   Location:
     -> Device Drivers
       -> Mailbox Hardware Support (MAILBOX [=y])
         -> Qualcomm Technologies, Inc. CPUCP mailbox driver (QCOM_CPUCP_MBOX [=n])

 Commit: 0e2a9a03106c (mailbox: Add support for QTI CPUCP mailbox controller)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>