[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_FSL_IMX8QM_HSIO [=n]
 Type  : tristate
 Defined at drivers/phy/freescale/Kconfig:38
   Prompt: Freescale i.MX8QM HSIO PHY
   Depends on: (ARCH_MXC [=y] && ARM64 [=y] || COMPILE_TEST [=n]) && OF [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> Freescale i.MX8QM HSIO PHY (PHY_FSL_IMX8QM_HSIO [=n])
 Selects: GENERIC_PHY [=y]

 Commit: 82c56b6dd24f (phy: freescale: imx8qm-hsio: Add i.MX8QM HSIO PHY driver support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>