acpi/x86: s2idle: Add ability to configure wakeup by AC adapter changes
According to https://learn.microsoft.com/en-us/windows-hardware/design/device-experiences/modern-standby-wake-sources#environmental-context-changes-1 the system MUST ALWAYS wake up the SOC when a power capable dock/charger is attached/detached to the device.

Currently, this causes issues on the original Ally and the Legion Go S. So, let's start nice and slow by quirking those two.

Add the ability for the LPS0 device to wake the system when AC
adapter state changes and add it to the delay quirk table.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>