[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_POE [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:2143
   Prompt: Permission Overlay Extension
   Location:
     -> Kernel Features
       -> ARMv8.9 architectural features
         -> Permission Overlay Extension (ARM64_POE [=y])
 Selects: ARCH_USES_HIGH_VMA_FLAGS [=y] && ARCH_HAS_PKEYS [=y]

 Commit: b9b9d72de32b (arm64: add Permission Overlay Extension Kconfig)

---

 Symbol: PTDUMP_STAGE2_DEBUGFS [=n]
 Type  : bool
 Defined at arch/arm64/kvm/Kconfig:69
   Prompt: Present the stage-2 pagetables to debugfs
   Depends on: VIRTUALIZATION [=y] && KVM [=y] && DEBUG_KERNEL [=y] && DEBUG_FS [=y] && GENERIC_PTDUMP [=y]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Kernel-based Virtual Machine (KVM) support (KVM [=y])
         -> Present the stage-2 pagetables to debugfs (PTDUMP_STAGE2_DEBUGFS [=n])
 Selects: PTDUMP_CORE [=y]

 Commit: 7c4f73548ed1 (KVM: arm64: Register ptdump with debugfs on guest creation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_PTDUMP_STAGE2_DEBUGFS enable for aarch64 debug.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>