[redhat] New configs in drivers/net/ethernet/realtek
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RTASE [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/realtek/Kconfig:123
   Prompt: Realtek Automotive Switch 9054/9068/9072/9075/9068/9071 PCIe Interface support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_REALTEK [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Realtek devices (NET_VENDOR_REALTEK [=y])
             -> Realtek Automotive Switch 9054/9068/9072/9075/9068/9071 PCIe Interface support (RTASE [=n])
 Selects: CRC32 [=y] && PAGE_POOL [=y]

 Commit: ad61903add56 (realtek: Update the Makefile and Kconfig in the realtek folder)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>