[redhat] New configs in drivers/net/wireless
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RTW89_8852BTE [=n]
 Type  : tristate
 Defined at drivers/net/wireless/realtek/rtw89/Kconfig:75
   Prompt: Realtek 8852BE-VT PCI wireless network (Wi-Fi 6) adapter
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_REALTEK [=y] && RTW89 [=m] && PCI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless LAN (WLAN [=y])
           -> Realtek devices (WLAN_VENDOR_REALTEK [=y])
             -> Realtek 802.11ax wireless chips support (RTW89 [=m])
               -> Realtek 8852BE-VT PCI wireless network (Wi-Fi 6) adapter (RTW89_8852BTE [=n])
 Selects: RTW89_CORE [=m] && RTW89_PCI [=m] && RTW89_8852BT [=n] && RTW89_8852B_COMMON [=m]

 Commit: b9cdbb06d4fc (wifi: rtw89: 8852bt: add 8852BE-VT to Makefile and Kconfig)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>