[redhat] New configs in drivers/perf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_NI [=n]
 Type  : tristate
 Defined at drivers/perf/Kconfig:51
   Prompt: Arm NI-700 PMU support
   Depends on: PERF_EVENTS [=y] && (ARM64 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Performance monitor support
         -> Arm NI-700 PMU support (ARM_NI [=n])

 Commit: 4d5a7680f2b4 (perf: Add driver for Arm NI-700 interconnect PMU)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_ARM_NI enabled as a module for aarch64.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>